Fix fractional cycle counts in various recipe trees
There were 40 cases throughout the default scenario where the number of recipe cycles required to produce enough of an item to make the given machine, tech or objective was fractional. Without this fix, all these cases would have lead to softlocks as, for example, an assembler wouldn't produce a single circuit to top it off as its recipe called for an output of 2, causing production to get stuck. This is fixed by making sure a machine can always run a whole number of times to produce just enough of a given item for the next step.